Fix negative bitshift error when no buckets are enabled
This is the same fix that was already applied to v1.3 branch.

Details below:

If a slug exists in the database but no buckets are enabled, we get back
an empty string. The empty string gets coreced into 0, then we subtact
1 to do the bitshift, and we end up with the negative bitshift error. In
php 5.x this is a silent error where the value ends up being
int(-9223372036854775808), but in php 7.1 it's a fatal error.

With this change when we reduce the array, if $index is 0 we just return
the $mask we have accumulated so far. There's no need to shift the 0
$index so we avoid it by returning early.

Below are some examples of what reduceToBitmask was returning:

Before:
```
When Test.a swivel_features = 1,2,3,4,5,6,7,8,9,10
reduceToBitmask() returned int(1023)

When Test.a swivel_features = ""
reduceToBitmask() returned int(-9223372036854775808)
```

Now:
```
When Test.a swivel_features = 1,2,3,4,5,6,7,8,9,10
reduceToBitmask() returns int(1023)

When Test.a swivel_features = ""
reduceToBitmask() returns int(0)
```

Signed-off-by: Rene Fragoso <ctrlrsf@gmail.com>